requirements.txt

# requirements.txt
langchain==0.1.0
langgraph==0.0.20
langchain-openai==0.0.5
langchain-community==0.0.10
transformers==4.36.0
torch==2.1.0
torchvision==0.16.0
sentence-transformers==2.2.2
chromadb==0.4.18
pymupdf==1.23.8
pillow==10.1.0
opencv-python==4.8.1.78
easyocr==1.7.0
plotly==5.17.0
streamlit==1.28.1
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
faiss-cpu==1.7.4
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
arxiv==1.4.8
matplotlib==3.8.2
seaborn==0.13.0
networkx==3.2.1
pydantic==2.5.0
typing-extensions==4.8.0
asyncio==3.4.3